+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - simplify release process
 - make challenge battles predictive
+- change default log level to info
+
+### Removed
+- PM2 dependency
 
 ## [2.1.11] - 2019-03-09